# go-admin-plus/api/openapi/openapi.json
{
  "components": {
    "schemas": {
      "ApiErrorEnvelope": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "data": {
            "nullable": true
          },
          "msg": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "msg"
        ],
        "type": "object"
      },
...
      "handler.Login": {
        "properties": {
          "code": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "uuid": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "password",
          "username",
          "uuid"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Authoritative Phase 1 HTTP wire contract for every go-admin host.",
    "title": "go-admin API",
    "version": "2.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/login": {
      "post": {
        "description": "获取token\nLoginHandler can be used by clients to get a jwt token.\nPayload needs to be json in the form of {\"username\": \"USERNAME\", \"password\": \"PASSWORD\"}.\nReply will be of the form {\"token\": \"TOKEN\"}.\ndev mode：It should be noted that all fields cannot be empty, and a value of 0 can be passed in addition to the account password\n注意：开发模式：需要注意全部字段不能为空，账号密码外可以传入0值",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/handler.Login"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "$ref": "#/components/schemas/ApiErrorEnvelope"
                    }
                  ]
                }
              }
            },
            "description": "{\"code\": 200, \"expire\": \"2019-08-07T12:45:48+08:00\", \"token\": \".eyJleHAiOjE1NjUxNTMxNDgsImlkIjoiYWRtaW4iLCJvcmlnX2lhdCI6MTU2NTE0OTU0OH0.-zvzHvbg0A\" }"
          }
        },
        "summary": "登陆",
        "tags": [
          "登陆"
        ]
      }
    },
    "/api/v1/logout": {
      "post": {
        "description": "获取token",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "string"
                    },
                    {
                      "$ref": "#/components/schemas/ApiErrorEnvelope"
                    }
                  ]
                }
              }
            },
            "description": "{\"code\": 200, \"msg\": \"成功退出系统\" }"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "退出登录"
      }
    },
...
  },
  "servers": [
    {
      "description": "Same-origin Web or host-provided Desktop endpoint",
      "url": "/"
    }
  ]
}